refactor: EntityKey 패키지 이동
